FEATURE: add additional fields in user list export

+++ app/models/export_csv.rb
@@ -10,10 +10,12 @@ class ExportCsv
   end
 
   def self.remove_old_exports
-    dir = Dir.new(ExportCsv.base_directory)
-    dir.each do |file|
-      if (File.mtime(File.join(ExportCsv.base_directory, file)) < 2.days.ago)
-        File.delete(File.join(ExportCsv.base_directory, file))
+    if Dir.exists?(ExportCsv.base_directory)
+      dir = Dir.new(ExportCsv.base_directory)
+      dir.each do |file|
+        if (File.mtime(File.join(ExportCsv.base_directory, file)) < 2.days.ago)
+          File.delete(File.join(ExportCsv.base_directory, file))
+        end
       end
     end
   end
